requirements.txt
Click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.5
redis==3.5.3
rq==1.7.0
Werkzeug==1.0.1
pyyaml==5.4.1
pandas==1.2.1
rpy2==3.4.2
biothings_client==0.2.3